Switches bootstrap to use the scss version instead of the compiled one
