+++ package.json
@@ -30,7 +30,8 @@
     "react": "^16.13.0",
     "react-bootstrap": "^1.0.0",
     "react-dom": "^16.13.0",
-    "react-markdown": "^4.3.1"
+    "react-markdown": "^4.3.1",
+    "sass": "^1.26.9"
   },
   "devDependencies": {
     "eslint": "^6.8.0",
