+++ src/components/index/header.jsx
@@ -1,0 +1,39 @@
+import Link from 'next/link';
+import Nav from 'react-bootstrap/Nav';
+import Navbar from 'react-bootstrap/Navbar';
+import Container from 'react-bootstrap/Container';
+
+const Header = () => (
+  <Navbar bg="dark" variant="dark" expand="md" sticky="top">
+    <Container>
+      <Link href="/" passHref>
+        <Navbar.Brand>Observatorio</Navbar.Brand>
+      </Link>
+      <Navbar.Toggle />
+      <Navbar.Collapse>
+        <Nav>
+          <Link href="/reports?category=registro-femicidios" passHref>
+            <Nav.Link>Femicidios</Nav.Link>
+          </Link>
+          <Link href="/reports" passHref>
+            <Nav.Link>Informes</Nav.Link>
+          </Link>
+          <Link href="/campaigns" passHref>
+            <Nav.Link>Campañas</Nav.Link>
+          </Link>
+          <Link href="/media-presence" passHref>
+            <Nav.Link>En Los Medios</Nav.Link>
+          </Link>
+          <Link href="/activities" passHref>
+            <Nav.Link>Actividades</Nav.Link>
+          </Link>
+          <Link href="/about" passHref>
+            <Nav.Link>Quiénes Somos</Nav.Link>
+          </Link>
+        </Nav>
+      </Navbar.Collapse>
+    </Container>
+  </Navbar>
+);
+
+export default Header;
