+++ src/styles/main.scss
@@ -1,3 +1,5 @@
+@use 'bootstrap/scss/bootstrap';
+
 img.content-image {
   max-height: 50vh;
   margin: auto;
